+++ sqelf/Cargo.toml
@@ -5,19 +5,22 @@ authors = ["Datalust"]
 edition = "2018"
 license = "Apache-2.0"
 
-[dependencies.futures-preview]
-version = "0.3.0-alpha.19"
-features = ["async-await"]
+[dependencies.futures]
+version = "0.3.0"
 
 [dependencies.pin-utils]
 version = "0.1.0-alpha.4"
 
 [dependencies.tokio]
-version = "0.2.0-alpha.6"
-features = ["signal"]
+version = "0.2.0"
+features = ["signal", "sync", "net", "time", "stream"]
+
+[dependencies.tokio-util]
+version = "0.2.0"
+features = ["codec", "udp"]
 
 [dependencies.bytes]
-version = "0.4"
+version = "0.5"
 
 [dependencies.libflate]
 version = "0.1"
